+++ docs/conf.py
@@ -51,7 +51,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'maxminddb'
-copyright = '2013-2018, MaxMind, Inc.'
+copyright = '2013-2019, MaxMind, Inc.'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the